Allow executing from any location
Previously, the script was expected to be executed from within the repo.
I created an alias and surprise, of course it stopped working. So I made
some tweaks that allow it to be executed from anywhere.

As part of that, I also change the -c and positional argument so that
they do different things. The positional allows you to keep your configs
in the ./configs directory of the repo, but the -c parameter allows you
to supply a config file from anywhere.